move the Windows 11 proxy style for menus into its own file
